requirements.txt
# Core dependencies
opencv-python==4.8.1.78
numpy==1.23.5
deepface==0.0.79
tensorflow==2.12.0

# Configuration management
pyyaml>=6.0

# Environment variables
python-dotenv>=1.0.0

# HTTP requests
requests>=2.31.0

# Advanced emotion recognition models (optional)
# Install based on your choice of detector
# For HSEmotion (recommended - fast and accurate):
hsemotion>=0.2.0

# For FER library (alternative):
# fer>=22.5.0

# For timm (required by HSEmotion):
timm>=0.9.0